+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 5.0.0 (unreleased)
 
+- Searches now use lazy loading (similar to Active Record)
 - Anchor regular expressions by default
 - Raise error when `search` called on relations
 - Raise `ArgumentError` (instead of warning) for invalid regular expression modifiers